refactor: Move sources into server/graphql dir
