+++ prisma/schema.prisma
@@ -7,6 +7,7 @@ generator client {
 
 generator fabbrica {
   provider    = "prisma-fabbrica"
+  output      = "../src/testing/factories/__generated__/fabbrica"
 }
 
 datasource db {
