+++ schema.graphql
@@ -1,4 +1,8 @@
 # This is auto generated file. Don't edit.
+schema {
+  query: Query
+}
+
 type Post {
   id: ID!
   title: String!
